requirements.txt
apache-airflow==2.9.0
boto3
python-dotenv